# specs/001-agent-core-spec/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Java Shannon Agent Orchestrator Core API
  version: 1.0.0
  description: Shannon 架构对齐的核心接口与事件流定义
servers:
  - url: /
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/policy/evaluate:
    post:
      summary: 策略评估
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyRequest'
      responses:
        '200':
          description: 评估成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PolicyDecision'
        '400':
          description: 参数错误，错误码：INVALID_REQUEST、TENANT_MISSING
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 策略拒绝，错误码：POLICY_DENIED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在，错误码：NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 服务不可用，错误码：CIRCUIT_OPEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/schedules:
    post:
      summary: 创建定时任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleSpec'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    put:
      summary: 更新定时任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleSpec'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    get:
      summary: 查询定时任务
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/schedules/{scheduleId}/pause:
    post:
      summary: 暂停定时任务
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 暂停成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/schedules/{scheduleId}/resume:
    post:
      summary: 恢复定时任务
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 恢复成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/schedules/{scheduleId}:
    delete:
      summary: 删除定时任务
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: object
        traceId:
          type: string
        requestId:
          type: string
    ErrorResponse:
      description: 错误码命名以 POLICY_DENIED、SANDBOX_DENIED 为准，别名 POLICY_DENY、SANDBOX_DENY 仅用于兼容说明。
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
        traceId:
          type: string
        requestId:
          type: string
          description: 请求唯一标识（来自请求上下文/请求头生成），用于端到端排查
...
    ScheduleSpec:
      type: object
      properties:
        scheduleId:
          type: string
        cron:
          type: string
        timezone:
          type: string
        status:
          type: string
        idempotencyKey:
          type: string
...
    PolicyRequest:
      type: object
      properties:
        policyId:
          type: string
        action:
          type: string
        resource:
          type: string
        input:
          type: object
    PolicyDecision:
      type: object
      properties:
        policyId:
          type: string
        decision:
          type: string
        reason:
          type: string
        evaluationId:
          type: string
        matchedRules:
          type: array
          items:
            type: string
...